重构three场景构件逻辑模型
